Extract Forecast structure to its own file
