+++ Source/Models/Forecast.swift
@@ -1,0 +1,10 @@
+//  Copyright © 2016 Gavan Chan. All rights reserved.
+
+import Foundation
+
+struct Forecast {
+  let city: City
+  let minimum: Temperature
+  let maximum: Temperature
+  let date: Date
+}
