+++ Source/Models/Weather.swift
@@ -1,6 +1,5 @@
 //  Copyright © 2016 Gavan Chan. All rights reserved.
 
-import Foundation
 
 typealias Temperature = Double
 typealias Percentage = Double
@@ -11,9 +10,3 @@ struct Weather {
   let cloudiness: Percentage
 }
 
-struct Forecast {
-  let city: City
-  let minimum: Temperature
-  let maximum: Temperature
-  let date: Date
-}
